Move getPeople function one level up to the parent component.
