+++ src/components/RequestUpdateForm.tsx
@@ -1,64 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useAppSelector } from "@/app/hooks";
-import { useNavigate } from "react-router-dom";
-import { useToast } from "@/hooks/use-toast";
 import RequestUpdateMealForm from "./forms/RequestUpdateMealForm";
 import RequestUpdateDonationServiceForm from "./forms/RequestUpdateDonationServiceForm";
-import createApi from "@/lib/api";
 
 function RequestUpdateForm() {
-    const navigate = useNavigate();
-    const { toast } = useToast();
-    const request = useAppSelector((state) => state.request.request)
-
-    const [people, setPeople] = useState([]);
-    // get current list of people in organization to search through
-    useEffect(() => {
-        const controller = new AbortController();
-        const api = createApi({ endpoint: "/people" });
-
-        async function getPeople() {
-            try {
-                const response = await api.get({ controller: controller });
-                const json = await response.json();
-
-                if (!response.ok) {
-                    if (response.status == 401) {
-                        toast({
-                            variant: "destructive",
-                            description: `${json.errors.detail}`
-                        });
-                        navigate("/login");
-                    } else if (response.status == 403) {
-                        toast({
-                            variant: "destructive",
-                            description: `${json.errors.detail}`
-                        });
-                        navigate("/");
-                    } else {
-                        toast({
-                            variant: "destructive",
-                            description: `${json.errors.detail}`
-                        });
-                    }
-                }
-                else {
-                    setPeople(json.data);
-                }
-            } catch (error) {
-                toast({
-                    variant: "destructive",
-                    description: `${error}`,
-                });
-            }
-        }
-
-        getPeople();
-
-        return () => {
-            controller.abort("Request Aborted");
-        }
-    }, [])
+    const request = useAppSelector((state) => state.request.request);
+    const people = useAppSelector((state) => state.people.people);
 
     return <>
         <div className="container mx-auto mb-5">
